refactor: navigation options
